Replace hardware SPI with a soft SPI
This is easier to debug with the help of a verilog flash emulator. There
is no emulation available for the hardware SPI so we can't debug in a
simulator.

Now the FreeRTOS app is able to run on hardware and a full system
simulation is also available running the same app.